requirements.txt
django==2.0.9
djangorestframework>=3.0.0
drf-chunked-upload